Style: Add local tests to format code
